requirements.txt
beautifulsoup4
lxml
google-cloud-texttospeech
google-cloud-storage
google-genai
firebase-admin
bleach
Flask
gunicorn
python-json-logger